test(examples): share streaming proxy lifecycle fixtures
Move raw backpressure cases into a focused suite and give the example and HTTP harness one owner. Configure upstream content count and terminal event independently while preserving all 47 cases and their assertions.
